Add path generator macro ROUTE-PATH and path prefix match feature.
Macro ROUTE-PATH calls a path generator to return a path string.

The router can be called with '(:list-routes "/foo") to list all
routes that match the specified prefix.

Also:
* Add declaim ftype statements
* Add docstrings
* Rename internal function DISPATCH-CLACK to DISPATCH
* Update README
* Add error ROUTER-COMPILATION-ERROR, base condition class for
  failures for COMPILE-ROUTER
* Add error NO-PATH-GENERATOR-ERROR, signalled when ROUTE-PATH
  ROUTE-PATH is called with an invalid route name
* Add error NO-ROUTE-FUNCTION-ERROR, signalled when the route
  symbol is missing a function value
* Rename INVALID-ROUTE-METADATA-ERROR to ROUTE-METADATA-ERROR
* Add struct FIND-ROUTE-RESULT, refactoring NODE and ROUTE-INFO
* Remove inline declaration from FIND-NODE
* Segment exports by type in package definition

NOTE: There is a subtle difference in behavior: route symbols
are coerced to keywords to check for uniqueness, so that
ROUTE-PATH can be called with a route symbol from packages loaded
prior to compiling the router. This means you cannot define routes
with the same symbol name defined in separate packages.